refactor: pull streaming file to component
